+++ LibraryApi/Models/Book.cs
@@ -7,11 +7,11 @@ namespace LibraryApi.Models;
 public class Book
 {
     [Key]
-    public long Id { get; set; }
+    public int Id { get; set; }
     public string Title { get; set; }
     public BookGenre Genre { get; set; }
     [ForeignKey("Author")]
-    public long AuthorId { get; set; }
+    public int AuthorId { get; set; }
     public Author Author { get; set; }
     public ICollection<Review> Reviews { get; set; }
 }
